+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## next
+## 4.0.0 (2 December 2017)
 
 - update UglifyJS to 3.2.0
 - remove Angular ngInject processing